+++ spec/www/spec/sqlite-version-test.js
@@ -161,10 +161,7 @@ var mytests = function() {
             expect(rs).toBeDefined();
             expect(rs.rows).toBeDefined();
             expect(rs.rows.length).toBe(1);
-            if (!isWebSql && !isWindows && isAndroid && isImpl2)
-              expect(rs.rows.item(0).page_size).toBe(4096); // CORRECT [androidDatabaseImplementation: 2]
-            else
-              expect(rs.rows.item(0).page_size).toBe(1024); // XXX TBD OLD VALUE USED IN THIS PLUGIN VERSION ref: litehelpers/Cordova-sqlite-storage#781
+            expect(rs.rows.item(0).page_size).toBe(4096);
 
             // Close (plugin only) & finish:
             (isWebSql) ? done() : db.close(done, done);
@@ -190,9 +187,8 @@ var mytests = function() {
             var resultRow = rs.rows.item(0);
             expect(resultRow).toBeDefined();
             expect(resultRow.cache_size).toBeDefined();
-            if (!isWebSql && isAndroid &&
-                (!isImpl2 ||
-                 (isImpl2 && (/Android [3-7]/.test(navigator.userAgent)))))
+            if (!isWebSql && isAndroid && isImpl2
+                && (/Android [3-7]/.test(navigator.userAgent)))
               expect(resultRow.cache_size).toBe(2000); // TBD OLD VALUE on Android (...)
             else
               expect(resultRow.cache_size).toBe(-2000); // NEW VALUE, otherwise